fix: handle native behaviour for off screen selection
When we press keyarrow down or up when we're off screen, the focused
element is scrolled into view first to the center of the scrollable
container.

This commit replicates that behaviour by first checking if the
selected element is in the viewport.
If not, we scroll into the center of the screen.