fix: handle chrome's behaviour for between lines
When selecting between lines, Chrome will set the position of the
caret to the start of the next line, instead of preserving the
horizontal.

When the margins between lines is significant, this can give a false
positive when we're searching for the correct range.

We need to check if this is a "suspicious" Chrome result and continue
our search if it is.
See CodeMirror's implementation for reference.
Note: the fix that I implemented here does not follow CodeMirror's
implementation directly.

https://github.com/codemirror/view/blob/main/src/cursor.ts#L230