fix: handle 2 dom rects on new line
At the edge of a line wrap, calling `range.getClientRects()`
returns 2 rectangles, 1 for the end of the current line and 1 for the
start of the next line.

`range.getBoundingClientRect()` returns the first rectangle, which looks
odd when the user is trying to scroll downwards.

This commit adds a check, and prefers returning the second rectangle
instead.
However, it doesn't entirely fix the issue, and we'll need to think of
a better way of handling this at a later date.