fix: handle cursor height on newlines
Cursor height used to follow the height of the container.

The height of the `span` on a line with text is lesser than the height
of the `p / h1 / h2 / etc.` for a line without any text.

As such, we should handle the height of the container, and adjust
the cursor's position slightly based on how much bigger the container
is compared to the calculated cursor height.

The cursor height is approximately 1.2x the height of the font.

Conveniently, this commit also fixes the issue of the cursor height
not updating when we change to headings
(as we're now calculating the cursor height from font size).