refactor: separate custom cursor and vim plugin
Cursor rendering and vim emulation are separate functions, and
don't have to be in the same plugin.

One additional step that's not yet refactored is to create custom
commands to be dispatched when the command mode changes,
such that parent components can subscribe to changes in the command mode
and render the custom cursor as a block cursor accordingly.